fix(orders): widen child ids, state the next id, digest id file keys
Type slOrderId and ptOrderId (Rust sl_order_id and pt_order_id) as the
order id is typed, so a family numbered from the ids a session hands out
is placed under them. Python no longer raises OverflowError for a wide id.

State next_valid_id once at connect on the Rust surface too, delivered
after the venue has named the working orders. A session that ends before
the id can be stated reports no error for it: the caller asked nothing.
order_id_floor() now reads the id next_valid_id states; next_shared_id()
keeps the floor a request can also carry.

Key the order-id file by the SHA-256 digest of the account, and rewrite a
file keyed by account numbers when a session opens it, keeping every
counter, the higher one where both keys hold one.

Leave an order's status unchanged on a report of a change on its way
(150=6), as a gateway does: a replacement in flight reads PendingSubmit
until the venue accepts it, and an order the venue revises itself keeps
its status, where both read PendingCancel before. An order whose state a
dropped connection lost is still read from that report.

Update the order documentation and the published test counts.

Fixes #37, fixes #38, fixes #39.
Refs #25.

+++ src/python/compat/class_orders.rs
@@ -249,7 +249,7 @@ pub struct Order {
     #[pyo3(get, set)]
     pub professional_customer: bool,
     #[pyo3(get, set)]
-    pub pt_order_id: i32,
+    pub pt_order_id: i64,
     #[pyo3(get, set)]
     pub pt_order_type: String,
     #[pyo3(get, set)]
@@ -305,7 +305,7 @@ pub struct Order {
     #[pyo3(get, set)]
     pub short_sale_slot: i32,
     #[pyo3(get, set)]
-    pub sl_order_id: i32,
+    pub sl_order_id: i64,
     #[pyo3(get, set)]
     pub sl_order_type: String,
     #[pyo3(get, set)]
@@ -624,7 +624,7 @@ impl Default for Order {
             post_only: false,
             post_to_ats: i32::MAX,
             professional_customer: false,
-            pt_order_id: i32::MAX,
+            pt_order_id: i32::MAX as i64,
             pt_order_type: String::new(),
             randomize_price: false,
             randomize_size: false,
@@ -650,7 +650,7 @@ impl Default for Order {
             settling_firm: String::new(),
             shareholder: String::new(),
             short_sale_slot: 0,
-            sl_order_id: i32::MAX,
+            sl_order_id: i32::MAX as i64,
             sl_order_type: String::new(),
             smart_combo_routing_params: ListField::new(),
             soft_dollar_tier: TierField::new(),
@@ -1978,7 +1978,7 @@ camel_aliases_copy! {
         get_post_only_alias set_post_only_alias postOnly post_only bool;
         get_post_to_ats_alias set_post_to_ats_alias postToAts post_to_ats i32;
         get_professional_customer_alias set_professional_customer_alias professionalCustomer professional_customer bool;
-        get_pt_order_id_alias set_pt_order_id_alias ptOrderId pt_order_id i32;
+        get_pt_order_id_alias set_pt_order_id_alias ptOrderId pt_order_id i64;
         get_randomize_price_alias set_randomize_price_alias randomizePrice randomize_price bool;
         get_randomize_size_alias set_randomize_size_alias randomizeSize randomize_size bool;
         get_ref_futures_con_id_alias set_ref_futures_con_id_alias refFuturesConId ref_futures_con_id i32;
@@ -1998,7 +1998,7 @@ camel_aliases_copy! {
         get_scale_subs_level_size_alias set_scale_subs_level_size_alias scaleSubsLevelSize scale_subs_level_size i32;
         get_seek_price_improvement_alias set_seek_price_improvement_alias seekPriceImprovement seek_price_improvement Option<bool>;
         get_short_sale_slot_alias set_short_sale_slot_alias shortSaleSlot short_sale_slot i32;
-        get_sl_order_id_alias set_sl_order_id_alias slOrderId sl_order_id i32;
+        get_sl_order_id_alias set_sl_order_id_alias slOrderId sl_order_id i64;
         get_starting_price_alias set_starting_price_alias startingPrice starting_price f64;
         get_stock_range_lower_alias set_stock_range_lower_alias stockRangeLower stock_range_lower f64;
         get_stock_range_upper_alias set_stock_range_upper_alias stockRangeUpper stock_range_upper f64;
